rjk-20010321.tar.gz
https://sourceforge.net/p/jos/mailman/message/2979253/

[jos-general] New release of RJK
From: Robert F. <ro...@27...> - 2001-03-21 11:36:53

Hi everybody

After about a year without any public releases of RJK (Robert's JOS
Kernel), I'm finally able to make a release.  The kernel now has support
for all the basic features that I think are needed to write a JVM on a
i386 processor.

The kernel has support for the following features:
* Atomic Operations
* IO Ports
* Threads
* Interrupts
* Virtual Memory
* Linear Memory
* Loading Modules
* Getting/Setting Time
* Timers
* Spinlock's
* Mutex's
* Printing Debug Messages

RJK goal in life is to be a minimal kernel, it should be able to load
it's self, the JVM image and any needed modules, it should then
implement all the required functions of the Kernel Interface.

RJK currently uses GRUB as it's boot loader, this allow it to load the
kernel and modules from ext2 or fat formated hard or floppy disks.  It
should even be possible to load the kernel and modules across a network.

The source and bootdisk image are available at
<URL:http://www.273k.net/jos/rjk/rjk-20010321.tar.gz> (24 KB) and
<URL:http://www.273k.net/jos/rjk/rjk-bootdisk-20010321.tar.gz> 71 KB.

The code is released under the GPL license and I plan to keep it out of
CVS for the moment until the source settles down a bit.

Enjoy.

Robert Fitzsimons
ro...@27...

Signed-off-by: Robert Fitzsimons <robfitz@273k.net>